refactor(order): reuse Common\Address, Identification, Phone in Order request layer
OrderAddressRequest, OrderIdentificationRequest, and OrderPhoneRequest were
duplicating structures already present in Resources\Common. Removed the three
redundant classes and updated OrderPayerRequest and OrderShipmentRequest to
use the shared Common counterparts.

Added JsonSerializable + toArray() to Common\Address, Identification, Phone
so they can serve as both response DTOs and request builders. Added the
missing `country` field to Common\Address (required by the Orders API) and
widened the city type to accept string|array|object|null to cover both
response (City object) and request (plain string) usage.
