+++ src/MercadoPago/Resources/Common/Address.php
@@ -2,17 +2,19 @@
 
 namespace MercadoPago\Resources\Common;
 
+use JsonSerializable;
 use MercadoPago\Serialization\Mapper;
 
 /**
  * Represents a physical address associated with a payer or shipment in the MercadoPago API.
  *
  * Used as a nested DTO within payer information, additional info, and shipment details
- * to describe street-level location data.
+ * to describe street-level location data. Also used as a typed request object when
+ * building order or payment request bodies.
  *
  * @see \MercadoPago\Resources\Payment\ReceiverAddress for the extended shipment address variant.
  */
-class Address
+class Address implements JsonSerializable
 {
     /** Maps nested objects to their corresponding DTO classes. */
     use Mapper;
@@ -44,8 +46,11 @@ class Address
     /** Apartment or unit identifier within a floor. */
     public ?string $apartment;
 
-    /** @var City|array|null City information associated with this address. */
-    public array|object|null $city;
+    /** Country name or code. Used in order request bodies. */
+    public ?string $country;
+
+    /** @var City|string|array|null City name (string in requests) or City object (in responses). */
+    public string|array|object|null $city;
 
     private $map = [
         "city" => "MercadoPago\Resources\Common\City"
@@ -57,5 +62,29 @@ class Address
     public function getMap(): array
     {
         return $this->map;
+    }
+
+    /** @return array<string,mixed> */
+    public function toArray(): array
+    {
+        $city = $this->city instanceof City ? $this->city->name ?? null : $this->city;
+        return array_filter([
+            "zip_code" => $this->zip_code,
+            "street_name" => $this->street_name,
+            "street_number" => $this->street_number,
+            "neighborhood" => $this->neighborhood,
+            "city" => $city,
+            "state" => $this->state,
+            "complement" => $this->complement,
+            "floor" => $this->floor,
+            "apartment" => $this->apartment,
+            "country" => $this->country,
+        ], fn ($v) => $v !== null);
+    }
+
+    /** @return array<string,mixed> */
+    public function jsonSerialize(): array
+    {
+        return $this->toArray();
     }
 }
